More robust save and trophy migration logic
